Re-add direct collection writes from admins
This is an escape hatch for admins with deep knowledge of the codebase
and data schemas to enable more straightforward manual frobbing of data
in production in urgent situations.  This should never be relied upon
for any product functionality.
