+++ imports/client/components/RTCDebugPage.tsx
@@ -6,11 +6,11 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, { useCallback, useState } from 'react';
 import Button from 'react-bootstrap/Button';
 import Table from 'react-bootstrap/Table';
+import isAdmin from '../../lib/is-admin';
 import CallParticipants from '../../lib/models/call_participants';
 import CallSignals from '../../lib/models/call_signals';
 import Profiles from '../../lib/models/profiles';
 import Puzzles from '../../lib/models/puzzles';
-import { isAdmin } from '../../lib/permission_stubs';
 import { CallParticipantType } from '../../lib/schemas/call_participants';
 import { CallSignalType } from '../../lib/schemas/call_signals';
 import { ProfileType } from '../../lib/schemas/profiles';
