+++ imports/client/components/SetupPage.tsx
@@ -13,10 +13,10 @@ import FormGroup from 'react-bootstrap/FormGroup';
 import FormLabel from 'react-bootstrap/FormLabel';
 import FormText from 'react-bootstrap/FormText';
 import Flags from '../../flags';
+import isAdmin from '../../lib/is-admin';
 import BlobMappings from '../../lib/models/blob_mappings';
 import DiscordCache from '../../lib/models/discord_cache';
 import Settings from '../../lib/models/settings';
-import { isAdmin } from '../../lib/permission_stubs';
 import { BlobMappingType } from '../../lib/schemas/blob_mapping';
 import { SettingType } from '../../lib/schemas/settings';
 import { DiscordGuildType } from '../discord';
